VoiceTranscriber/requirements_transcriber.txt
# transformers==4.41.1
# faster-whisper==1.0.1
# pyannote.audio==3.1.1
# librosa==0.10.1
# pydub==0.25.1

transformers
faster-whisper
pyannote.audio
librosa
pydub